+++ Assets/Core/Runtime/Navigation/ViewOption.cs
@@ -1,8 +1,0 @@
-﻿using UnityEngine;
-
-namespace KKSFramework.Navigation
-{
-    public class ViewOption : MonoBehaviour
-    {
-    }
-}