+++ Assets/Core/Runtime/Scripts/Navigation/AreaBase.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace KKSFramework.Navigation
+{
+    public abstract class AreaBase<T> : MonoBehaviour
+    {
+        public abstract void SetArea (T areaData);
+    }
+}