+++ Assets/Core/Runtime/Scripts/Navigation/ViewOption.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine;
+
+namespace KKSFramework.Navigation
+{
+    public class ViewOption : MonoBehaviour
+    {
+    }
+}