+++ Assets/Core/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2020-08-30
+### Added
+* Add [NavigationComponent] prefab from Base Project to Core.
+
+### Change
+* root [Canvas], [EventSystem] declare in [NavigationComponent].
+* Change to binding system [NavigationComponent] reference.
+* Separate Core folder Scripts and Prefab.  
+
 ## [1.2.0] - 2020-08-30 (Property Binding System : DataBind)
 ### Added
 * Add property and properties bind system on data bind.